+++ src/main/java/com/beyond/easycheck/user/application/service/admin/AdminReadUseCase.java
@@ -1,0 +1,4 @@
+package com.beyond.easycheck.user.application.service.admin;
+
+public interface AdminReadUseCase {
+}
