+++ src/main/java/com/beyond/easycheck/user/application/service/user/UserOperationUseCase.java
@@ -1,9 +1,7 @@
-package com.beyond.easycheck.user.application.service;
+package com.beyond.easycheck.user.application.service.user;
 
 import com.beyond.easycheck.corporate.ui.requestbody.CorporateCreateRequest;
-import com.beyond.easycheck.user.application.domain.UserStatus;
-import com.beyond.easycheck.user.application.service.UserReadUseCase.FindJwtResult;
-import com.beyond.easycheck.user.application.service.UserReadUseCase.FindUserResult;
+import com.beyond.easycheck.user.application.service.user.UserReadUseCase.FindJwtResult;
 import org.springframework.web.multipart.MultipartFile;
 
 public interface UserOperationUseCase {
@@ -11,8 +9,6 @@ public interface UserOperationUseCase {
     void registerUser(UserRegisterCommand command);
 
     void registerCorporateUser(UserRegisterCommand command, CorporateCreateRequest corporateCreateRequest, MultipartFile verificationFilesZip);
-
-    FindUserResult updateUserStatus(UserStatusUpdateCommand command);
 
     FindJwtResult login(UserLoginCommand command);
 
@@ -45,12 +41,6 @@ public interface UserOperationUseCase {
     ) {
     }
 
-    record UserStatusUpdateCommand(
-            Long userId,
-            UserStatus status
-    ) {
-
-    }
     record UserLogoutCommand(
             String accessToken,
             Long userId
