+++ src/main/java/com/beyond/easycheck/user/ui/view/UserView.java
@@ -4,7 +4,7 @@ import com.fasterxml.jackson.annotation.JsonInclude;
 import lombok.Getter;
 import lombok.ToString;
 
-import static com.beyond.easycheck.user.application.service.UserReadUseCase.FindUserResult;
+import static com.beyond.easycheck.user.application.service.user.UserReadUseCase.FindUserResult;
 
 @Getter
 @ToString
